pyproject.toml
[tool.poetry]
name = "sc2city"
version = "0.1.0"
description = ""
authors = ["savelas81 <savelas81@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
burnysc2 = "^6.4.0"
aiohttp = "^3.8.6"
black = "^23.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
known_first_party = "utils,game_objects,managers,config,placement_extractor,build_order_extractor,third_party_code"